fake-news-detector/requirements.txt
# Core Flask dependencies
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1

# Authentication and security
bcrypt==4.1.2
authlib==1.3.0

# AI/ML dependencies
groq>=1.1.1
sentence-transformers>=2.7.0

# Web scraping and HTTP
requests==2.31.0
newspaper3k==0.2.8

# Additional utilities
beautifulsoup4==4.12.2